Add files via upload
Interactive interface to perform basic functions such as addition, subtraction, division, and
multiplication. The Calculator is done by using HTML , CSS , JAVASCRIPT.A display screen to display the user’s input and give the results. The grid
system in CSS is for the alignments of buttons on the calculator.  Other tools by using JAVASCRIPT to make fully functional.